+++ ios/Classes/Annotation/Options.swift
@@ -2,7 +2,7 @@
 class Options<G: Geometry> {
 
     // Work around because Swift does not support generic protocols (or abstract classes).
-    func build(id: Float) -> Feature<G>? {
+    func build(id: UInt64) -> Feature<G>? {
         assert(false, "This method must be overriden by the subclass")
     }
 }
